+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Automated data updates are no longer made on a constant basis; rather, data is updated only when changes are needed because of a process
 * The layout of the `/about` embed response was slightly updated.
 * `/pronouns` only allows adding/removing one pronoun role at once.
+* `/forums`, `/obb`, `/gallery`, and `/exchange` were simplified into `/link`.
 
 ### Removed
 * The exalt/unexalt commands were removed.